Fix #188 Fix #189 Fix #151 Fix #177 Fix #155 Fix #132 Fix #174 Fix #170
Appart from the issus above, in this commit I have also fixed:
publisher, alt in images to avoid broken images when no connection is
available. Fixed in save metadata that the publisher is also saved.
I fixed issues with the consistency in sections and back to toc titles.

The toc will be now dynamic