Parsing external owl entities (direct/indirect imports)
The xslt parser cannot extract metadata from imported ontologies when
`-includeImportedOntologies` is not enabled. This is important because
sometimes you do not want to document the imported ontologies but you
still want to include metadata from the imported ontologies.

This change allows to tag owl entities that are not part of the
main ontology (i.e. direct/indirect imports) as an external property.
With the owlapi an ExternalPropertyParser class is added which looks for the
external property superscripts in the html generated by LODE and tries to find
the owl entity based on the IRI.

The xslt extraction sheet was modified to add the `external property` superscript
and html class `type-ep`.