Added local import feature
This change allows to use local imports with the
AutoIRIMapper from the owl api. The main use case
is to load imports locally for optimization of
loading time.

To enable this option use "-import" with a directory path.
The option can be used multiple times in the same command line.