fix(extraction): Do not force lower-case if URI do not have label
In the case the URI do not have any label, a generic one is generated
regarding the URI fragment. It was forced to lower case, which leads to
a miscomprehension for class label. It is commonly accepted that classes
label starts with a uppercase letter.

For example:
```turtle
:MyClass a owl:Class;
         owl:equivalentClass schema:Person.
```
leaded to a result in the doc like:
"MyClass is equivalent to person"
When we want :
"MyClass is equivalent to Person"