include SKOS alternative labels in documentation
Motivation:

From the SKOS primer:
> The skos:altLabel property makes it possible to assign an alternative
> lexical label to a concept. This is especially helpful when assigning
> labels beyond the one that is preferred for the concept, for instance
> when synonyms need to be represented:

Such properties are currently not included in the Widoco output.  The
absence of altLabel in the documentation creates a barrier to people who
are most familiar with the concept in terms of a synonym: a person might
be unsure if the term is the desired one if the synonym isn't listed, or
might search (on the webpage) for a term using the synonym, etc.

Modification:

The XSLT is updated to generate a new section in the metadata part of an
entity's description.  This new section is only present if the entity
has at least one `skos:altLabel` assertion, otherwise no additional
output is generated.  If the entity has at least one `skos:altLabel`
assertion then a section header is included followed by a list of the
`skos:altLabel` assertion values, using HTML's description list
(`dl`/`dt`/`dd`).

This new section is labelled `Alternative label` in English locale
output.  Auto-translated equivalent terms has been added for the other
supported languages.

Result:

The generated webpage(s) now include information from `skos:altLabel`
assertions.

Closes: #779