Identify non-release builds as SNAPSHOT
Motivation:

Current builds contain a complete version, which could lead to
confusion.

Modification:

Adopt maven's SNAPSHOT support.  This is a feature where builds that are
not from a release are given a SNAPSHOT label by appending `-SNAPSHOT`
to the version string.  The maven `release` plugin handles updating the
version and tagging of releases.

Result:

Less confusion over intermediate builds.

Signed-off-by: Paul Millar <paul.millar@desy.de>